+++ reservation/src/main/java/com/bootcamp/reservation_api/model/Reservation.java
@@ -1,9 +1,8 @@
-package com.bootcamp.reservation.model;
+package com.bootcamp.reservation_api.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Value;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
