+++ reservation/src/main/java/com/bootcamp/reservation_api/repository/FakeReservationRepository.java
@@ -1,6 +1,6 @@
-package com.bootcamp.reservation.repository;
+package com.bootcamp.reservation_api.repository;
 
-import com.bootcamp.reservation.model.Reservation;
+import com.bootcamp.reservation_api.model.Reservation;
 import lombok.AllArgsConstructor;
 
 import java.util.*;
