+++ reservation/src/main/java/com/bootcamp/reservation_boot/ReservationApplication.java
@@ -1,4 +1,4 @@
-package com.bootcamp.reservation;
+package com.bootcamp.reservation_boot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
