+++ reservation/src/test/java/com/bootcamp/reservation_api/ReservationApplicationTests.java
@@ -1,4 +1,4 @@
-package com.bootcamp.reservation;
+package com.bootcamp.reservation_api;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
